requirements.txt
torch>=1.11
torchvision
scikit-learn
pandas
tqdm
python-dotenv